refactor: Middlewares handled in api folder. Called in factory. Global exceptions and responses handling.
